Small screen navbar fix with style tweaks (#30)
* navbar: fix social icons hanging below header on small screens

As a consequence of the large font sizes the social icons were
hanging below the header on smaller screens (around 400px width,
approx.).

Fix that by adapting the font size for smaller screens. This should
fix the hanging issue at least for screen widths approx. upto 320px.

* style-tweaks: make the text in the header bold

This is to improve the readability of it. It's hard to read with the
light text. Particularly after the reduction of font size in smaller
screens.

* style-tweaks: reduce font size of blog post titles on small screens

They appear very large on small screens.

So, reduce the size a bit on small screens.